Fix weapon range parsing and display
- Add support for range inside weapon parentheses: Weapon (6", A1)
  Previously only supported: 24" Weapon (A1)
- Fix text formatter outputting u8 as char instead of int for range

+++ include/export/text_formatter.hpp
@@ -144,7 +144,7 @@ private:
                 for (const auto& w : model.weapons) {
                     out << "        - " << w.name << " (";
                     if (w.range > 0) {
-                        out << w.range << "\", ";
+                        out << static_cast<int>(w.range) << "\", ";
                     } else {
                         out << "Melee, ";
                     }
